split InprudentTweetActivity
