+++ app/src/main/AndroidManifest.xml
@@ -1,34 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="jp.ac.titech.itpro.pbl.burning_pro">
+	package="jp.ac.titech.itpro.pbl.burning_pro">
 
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+	<application
+		android:allowBackup="true"
+		android:icon="@mipmap/ic_launcher"
+		android:label="@string/app_name"
+		android:roundIcon="@mipmap/ic_launcher_round"
+		android:supportsRtl="true"
+		android:theme="@style/AppTheme">
+		<activity android:name=".MainActivity">
+			<intent-filter>
+				<action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
+				<category android:name="android.intent.category.LAUNCHER" />
 
-                <!--
+				<!--
                 intent filter for twitter
                 see: https://developer.android.com/guide/components/intents-common?hl=ja#ViewUrl
                 -->
-                <action android:name="android.intent.action.VIEW" />
-                <!-- Include the host attribute if you want your app to respond
-                     only to URLs with your app's domain. -->
-                <data android:scheme="http" android:host="www.twitter.com" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <!-- The BROWSABLE category is required to get links from web pages. -->
-                <category android:name="android.intent.category.BROWSABLE" />
+				<action android:name="android.intent.action.VIEW" />
+				<!--
+                     Include the host attribute if you want your app to respond
+                     only to URLs with your app's domain.
+                -->
+				<data
+					android:host="www.twitter.com"
+					android:scheme="http" />
 
-            </intent-filter>
-        </activity>
-    </application>
+				<category android:name="android.intent.category.DEFAULT" />
+				<!-- The BROWSABLE category is required to get links from web pages. -->
+				<category android:name="android.intent.category.BROWSABLE" />
+			</intent-filter>
+		</activity>
+		<activity android:name=".ImprudentTweetActivity"></activity>
+	</application>
 
 </manifest>
