+++ app/src/main/res/layout/activity_imprudent_tweet.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	tools:context=".ImprudentTweetActivity">
+
+	<Button
+		android:id = "@+id/tweetTest1"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:text="Test1"
+		android:onClick="tweetTest"
+		app:layout_constraintBottom_toBottomOf="@id/tweetTest2"
+		app:layout_constraintLeft_toLeftOf="parent"
+		app:layout_constraintRight_toRightOf="parent"
+		app:layout_constraintTop_toTopOf="parent" />
+
+	<Button
+		android:id = "@+id/tweetTest2"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:text="Test2"
+		android:onClick="tweetTest2"
+		app:layout_constraintBottom_toBottomOf="@id/tweetTest3"
+		app:layout_constraintLeft_toLeftOf="parent"
+		app:layout_constraintRight_toRightOf="parent"
+		app:layout_constraintTop_toTopOf="@id/tweetTest1" />
+
+	<Button
+		android:id = "@+id/tweetTest3"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:text="Test3"
+		android:onClick="tweetTest3"
+		app:layout_constraintBottom_toBottomOf="@id/generateTest"
+		app:layout_constraintLeft_toLeftOf="parent"
+		app:layout_constraintRight_toRightOf="parent"
+		app:layout_constraintTop_toTopOf="@id/tweetTest2" />
+
+	<Button
+		android:id = "@+id/generateTest"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:text="generate_Test"
+		android:onClick="generateTest"
+		app:layout_constraintBottom_toBottomOf="parent"
+		app:layout_constraintLeft_toLeftOf="parent"
+		app:layout_constraintRight_toRightOf="parent"
+		app:layout_constraintTop_toTopOf="@id/tweetTest3" />
+
+	<TextView
+		android:id = "@+id/ImprudenceText"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:text=""
+		app:layout_constraintBottom_toBottomOf="parent"
+		app:layout_constraintLeft_toLeftOf="parent"
+		app:layout_constraintRight_toRightOf="parent"
+		app:layout_constraintTop_toTopOf="@id/generateTest" />
+
+</android.support.constraint.ConstraintLayout>
